finance_project/requirements.txt
django
djangorestframework
psycopg2
redis
django-redis
django-environ
coverage
djangorestframework-simplejwt
drf-spectacular